[WIP] Add a Fuchsia VM runner and run a single command.
Since libFuzzer is built into Fuchsia, and thus Fuchsia can do *most* of
the fuzzing legwork on-target, we actually only need ClusterFuzz to
serve as a thin "wrapper" for booting and managing Fuchsia VMs, and
viewing the resultant crashes, in an automated way.

This is the first step toward that universe.

(It's mostly me realizing I was way overcomplicating things earlier,
oops.)

This changeset adds a new runner to libfuzzer.py,
FuchsiaQemuLibFuzzerRunner, which (1) downloads all the bits needed to
run Fuchsia from a GCS bucket (in particular: QEMU itself, a QEMU kernel,
an initrd, a qcow, and a one-off ssh_config specific to that Fuchsia VM),
(2) boots a VM, and (3) attempts to run a simple "echo hello"-style test
on that VM via SSH.

Stuff that will probably get fixed soon:

* A more elegant way of waiting for the QEMU server to start.

Stuff that'll come shortly after this changeset:

* Make a job per fuzzer.  (Find a way to automate this?)

* Extract crashes from our QEMU targets.

Stuff that will get fixed eventually but less soon:

* Downloading all this stuff from a GCS bucket is not super-elegant.
Chromium tests, for instance, build Fuchsia images from scratch when
running tests.  However, given that it's quite simple to build a version
of QEMU which has *all* available fuzzers built in (`fx set --fuzz-with
asan`), some kind of rotation where we upload the latest build every so
often (or set up an automated way to do so) seems reasonable.

Long-term we'd like to GCE VMs, but those only recently became
available, so let's hold off on them for a bit.

* Running multiple QEMU instances on one bot.  Running multiple tasks
per bot actually only recently became a thing in Clusterfuzz, but we may
as well take advantage of it.  If we're using QEMU, it'll involve
dynamically creating hostfwd ports for each QEMU instance.  If we're
using GCE, it'll involve something fancier.

TODOs BEFORE SUBMITTING:

* Need to actually make a bucket where we dump this stuff
* Need to test stuff end-to-end (are there automated tests for the runners?)