This one does NOT work (that is, seems to repro the failure?)
It must be someting weird that happens when you try to use the files
downloaded from the bucket.

Either it silently chokes on a chmod (would be weird, since whoami is
flowerhack) or it um.  Actually I'm confused.  I dunno

Works when run from command line.

At least this is progress.